requirements.txt
# Essential build tools - MUST be first
setuptools>=70.0.0
wheel>=0.41.0
pip>=24.0

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication & Security  
bcrypt==4.1.2
PyJWT==2.8.0

# Basic utilities
requests==2.31.0
pydantic==2.5.3
jinja2==3.1.2
click==8.1.7

# Image processing basics
Pillow==10.0.1
numpy==1.24.3

# Computer Vision - only if needed
opencv-python-headless==4.8.1.78

# ML libraries - simplified for better compatibility
# deepface==0.0.79
# tensorflow-cpu==2.15.0
# tf-keras==2.15.0
# retinaface==0.0.17
